refactor: add lossy_reader mod
