+++ src/lib.rs
@@ -1,2 +1,3 @@
 pub(crate) mod core;
+pub mod lossy_reader;
 pub use crate::core::*;
